+++ packages/graphback-runtime/src/data/GraphbackDataProvider.ts
@@ -49,15 +49,6 @@ export interface GraphbackDataProvider<Type = any, GraphbackContext = any> {
   delete(data: Type, context?: GraphbackContext): Promise<Type>;
 
   /**
-   * Implementation for finding all objects
-   *
-   * @param name name of the object to create
-   * @param id id of the object
-   * @param context context object passed from graphql or rest layer
-   * @deprecated Please use `findBy`
-   */
-  findAll(page?: GraphbackPage, context?: GraphbackContext): Promise<Type[]>;
-  /**
    * Implementation for finding a single unique object
    *
    * @param filter filter by unique attriburtes