+++ templates/ts-react-apollo-client/src/App.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import './App.css';
-import { useFindAllNotesQuery } from './generated-types';
+import { useFindNotesQuery } from './generated-types';
 import CreateNote from './components/notes/CreateNote';
 import OneNote from './components/notes/OneNote';
 
 const App: React.FC = () => {
-  const allNotes = useFindAllNotesQuery();
+  const allNotes = useFindNotesQuery();
   allNotes.startPolling(2000);
-  console.log(allNotes.data?.findAllNotes)
+  console.log(allNotes.data?.findNotes)
 
   return (
     <div>
@@ -15,7 +15,7 @@ const App: React.FC = () => {
       <ul>
         {
           // TODO fix typings
-          allNotes.data && allNotes.data.findAllNotes.map((note: any) => (
+          allNotes.data && allNotes.data.findNotes.map((note: any) => (
             <OneNote key={note.id} id={note.id} title={note.title} description={note.description} comments={note.comments}></OneNote>
           ))
         }